MR : Fix | controller => deploy-test

See merge request s09-webmobile3-sub2/S09P12A304!150

+++ BackEnd/ggong/src/main/java/com/a304/ggong/dto/response/UserResponse.java
@@ -15,9 +15,9 @@ public class UserResponse {
     private String gender;
     private String favoriteCigarette;
     private String QR;
-    private Long points;
+    private Integer points;
 
-    public UserResponse(User user, Long points) {
+    public UserResponse(User user, Integer points) {
         this.userNo = user.getUserNo();
         this.email = user.getEmail();
         this.nickname = user.getName();
